Split out camera to make it easier to understand
